- Merged KMetaData, Konto, and KNepomuk into one nepomuk folder with two libs:   1. libnepomuk now contains KMetaData and Konto, only using one single namespace "Nepomuk"   2. libnepomuk-middleware is now what knepomuk was before and has the namespace "Nepomuk::Middleware"
This makes the design much cleaner and less confusing since we have one name for everything: Nepomuk.
I was not happy with Braid and we did not find another name so I stick to Nepomuk now. In the end people
now already now this stuff as Nepomuk and it makes the project happy. ;)

CCMAIL: nepomuk-kde.semanticdesktop.org

svn path=/trunk/KDE/kdelibs/; revision=674110