It is about time I commit this to trunk: For a long time I have not been using inotify since the default user watch limit was too low on most systems. With new kernels and fsnotify or a higher user watch limit (as already done by most distris) the limit should be no problem.
However, with a lot of subfolders simply installing all the watches can take a long time (up to 10 minutes
on my system with all the KDE code trees). Thus, the IO priority is also lowered now.
We need to watch each an every folder separately since we need to catch file moves.

With this commit the biggest part of use cases should be covered. Since we do not watch the whole folder 
tree (only the home dir + all the strigi index folders) we will never be able to update an annotation on 
a moved file in /tmp or /var or whatever the user might need.

The next commit will adjust the strigi service to the new filewatch.

svn path=/trunk/KDE/kdebase/runtime/; revision=1104720