nepomuk: Parenthesize default argument expression for `mappings' (build fix).
This should fix the build on older compilers (at least gcc <= 4.3.3).

I suspect this is GCC bug 57 [1][2][3], which depends on a final
answer from the C++ committee. For some reason, g++ has changed its
default behaviour in a later release in a way that made this kind of
construction work.

CCMAIL: handa.vish@gmail.com

[1] http://gcc.gnu.org/bugzilla/show_bug.cgi?id=57
[2] http://gcc.gnu.org/bugzilla/show_bug.cgi?id=18688
[3] http://www.open-std.org/JTC1/SC22/WG21/docs/cwg_active.html#325