Faster Identification when minScore == 1.0
The identifiation process is very slow in the case when minScore == 1.0. This is because findMatch calculates the score for each property matched and then prunes the resources whose score falls below the minScore. This all cannot be done in one sparql query.

However, when minScore == 1.0. The entire identification process gets reduced to one query. Making it a gazillion times faster!