Spawn the Strigi Indexer in a separate process
This creates an executable "nepomukstrigiindexer" that is responsible for
indexing a given file. It calls the required strigi indexer, and pushes the
changes into Nepomuk.

This split has been done so that faulty strigi plugins do not make the
nepomuk strigi service crash.

Some code is still shared between this new process and the service. This should
be fixed in future commits.