Fixed a major memory leak.
Actually this was not really a leak. The queue of outgoing DBus messages
just filled up and was never cleared. I am not entirely sure why but
the thread emitting the signals (the original signals which trigger
emitting of the final DBus signals) needs event handling. In this case
it was the DataManagementCommand. I decided to simply handle events
using QEventLoop::processEvents. While this might not be very clean
an actual execution of an event loop and a queued execution of the
command in question would have the same effect since all commands are
executed sync anyway.

Thus, while this fixes the memory leak it is not very satisfying as I
do not understand why. :(