Fixed full text scoring.
Full text scoring now works whenever bif:contains can be used. Scoring
for REGEX filters is of course not possible.

I also fixed all unit tests and optimized the full text search excerpt
handling (we do not want to highlight search terms that should not be
part of the result, ie. those that are prefixed with "NOT").