Do not keep app as maintainer of metadata of removed resources.
In removeDataByApplication if an app removes a resource which has data
created by another app the resource's metadata is preserved.
With this change it is made sure that the calling app is not left as
the maintaining app of this metadata.
Thus, querying for resources maintained by the calling app will never
return the removed resources.

This fixes the testRemoveDataByApplication_nieUrl unit test.