Removed all redundant checks in the indexer.
Now that DMS does perform all the necessary checks and conversions
there is no need for that in the indexer anymore. This makes the code
simpler and easier to maintain.

There is only one thing I had to introduce (not implemented yet though):
the LazyCardinalities flag for the storeResources method which will
make it simply drop and value that would violate a cardinality
restriction.