New and improved describeResources method.
The DataManagementModel's describeResources now uses flags instead of
the boolean parameter. The supported flags are
- ExcludeDiscardableData - does exactly what it says: it excludes all
  data from discardable graphs.
- ExcludeRelatedResources - excludes related resources except sub-
  resources. By default all related resources are returned with their
  identifying properties. If this flag is set only sub-resources are
  included.

As mentioned before sub-resources are now always included. A sub-
resource is defined as being part of the super-resource and not making
any sense without it. Thus, excluding sub-resources is pointless.

The method also includes a third parameter: targetParties. This is not
implemented yet but will be used to specify the nao:Partys that the
data should be sent to. describeResources will then filter out data
that is not public to all of the parties.

Many situations are covered in the unit tests although probably not all.

One Virtuoso issue remains: all queries have to be executed for each
resource individually. I already asked on the Virtuoso mailing list[1]
about this issue. To me it seems like a bug.

[1] http://sourceforge.net/mailarchive/forum.php?thread_name=4E36D5ED.9020904%40kde.org&forum_name=virtuoso-users