Now a "real" hacky "fix" for the problem with wide characters in queries.
My last "fix" just used bif:charset_recode on the search terms. This,
however is no real solution as it makes the query fail if the terms
do not contain any wide character. Thus, the problem was simply
inverted.
How I do it the brute-force way: I simply truncate the terms at the
first wide char and use the rest. The search excerpts have the same
quality as with the bif:charset_recode hack but it works for wide and
non-wide terms.

CCBUG: 281136