nepomukserver does not need to be a KApplication.
The only reason to use KApplication or rather KUniqueApplication before
was to ensure that only one server is running. However, we do not need
to be able to attach to a running server or provide it with command
line arguments. Thus, a simple DBus service test is perfectly sufficient
and allows us to scale the server down to a QCoreApplication.

That way we also do not expose the KApplication DBus interface. This is
important since we do not want anyone to call the quit method without
the server being able to gracefully shut down.