Normalize optical media labels for usage as identifiers.
Sadly we have to normalize the labels by removing spaces and converting
all to lower-case since we use them as the host in our URLs. This means
we have to bend to the rules of QUrl which does not handle spaces well
at all (the host is simply dropped) and converts all to lower case.