Handle kext:altUrl as a special case.
We use kext:altUrl for local symlink handling. To that end it needs to
be treated as a special case just like nie:url since we do not want to
create a new resource but directly use the URL as value for the property.

This is true for add/setProperty and removeProperty, ie. all methods
that have a parameter which takes property values.