Disable query excerpts by default
If people want them, they should enable them. There is no advantage
of providing them by default. Specially since no one seems to be using
them. It also complicates the queries when you're trying to debug stuff.

Another advantage of not having search excerpts is that the same result
is not returned multiple times when doing a literal search.

This breaks a lot of the existing tests. Those tests need to be rewritten.

REVIEW: 105065