Do not create the DiffGenerator and SyncManager
They aren't used, and they just consume memory AND the DiffGenerator
connects to the nepomukserver ( cause of the ResourceManager ), which
results in the creation of a new thread in nepomukstorage. This new
thread is just lying there doing absolutely nothing.

Ideally, most of this code should just be thrown away.