Allow synchronous execution of queries
This is done by returning a Nepomuk::Query::ResultIterator, and
allows the user to control the execution of the query with the
iterator.

This lets us move away from Soprano. Plus, this avoids
code duplication as this code is being duplicated in both the
query service and the kioslave.

REVIEW: 105287