Remove special handling for rdf:type
This includes storing the maintype when loading all the properties. This
simplifies the code base, and actually improves performace.

Because the type deduction is done during load time, the relevant types
needed to be loaded into memory by Types::EntityPrivate, this increases
the load time for each resource by an order of 5x. This extra load
eventually goes down when all the required types are loaded, but it's
there.

Also, the additional cost of infering the type each time doesn't seem to
have any performance loss. Mainly cause everything is already in memory.

Another reason for this patch is the planned refactoring of the class
internals to use shared memory. Every extra variable in ResourceData
adds an overhead.

REVIEW: 105517