Add static Tag::allTags()
This was removed during the port to Nepomuk2, since we did not feel it
was a good idea having a static function which could block for quite
some time when the tag list gets pretty big.

But it seems as though loads of people are using it, and using the
asynchronous version (ResourceTypeTerm), would require significantly
restructuring of their code.