Add Resource::symbols()
They still return a string list instead of a QList<Resource>, but that's
only cause they are easier to use this way. And this way we can avoid
duplicates, cause I doubt the client will actually check for duplicates.
Specially since checking for duplicates involves writing a sparql query.

REVIEW: 105570