Add an environment variable NEPOMUK_TESTLIB_RUNNING
This variable is checked before running any of the tests. If the value
of the variable has not been set, then the tests are aborted.

At the end of every test the entire Nepomuk repository is purged. This
is the behaviour one wants for tests, but not for the production system.