Ractor ResourceData and ResourceManager
Improve the concepts of kickoff lists and make the code a lot easier to
understand. Most of the "magic" of the kickoff lists has been removed to
make it more explicit, and easier to understand.

Instead of having one kickOff list for everything. We now have 3 lists -
* One for initialized data. Maps uri -> ResourceData*
* One for nao:identifiers. Maps nao:identifier -> ResourceData*
* One for nie:urls. Maps nie:urls -> ResourceData*

When constructing a ResourceData, the ResourceManager checks the kind
QUrl it has been given - a nepomukuri, a nie:url or nao:identifier.
Based on the results it passes the arguemnts to the ResourceData
constrcutor.

This change was only possible cause with the DMS, we have made it clear
that all resource uri's will be of the form 'nepomuk:/'. This makes the
logic a lot simpler.

REVIEW: 105622