Introduce fall-back indexing
If the strigi indexer fails to index the file, we index it with this new
"simple indexer" which just sets the basic properties like url,
filename and the mimetype.

This way, file which weren't indexed, will at least still appear in some
of the search results.

The nie:lastModifed for these files has not been set, so the
indexscheduler will still find these files, and try to index them again.

REVIEW: 105730