TestLib: Cleanup the custon dbus session code
dbus-launch was called with --exit-with-session, which creates a
persistent "babysitter" process that watches stdin for SIGHUP, and tries
to connect to the X-Server. If this process gets a SIGHUP or looses its
conection to the X-server, it kills the message bus-daemon.

We do not require any of this overhead, cause we do not actually run a
separate x-server for the tests.

The Dbus code is now split into two files -
* dbus-session-begin.sh
* dbus-session-end.sh - Manually kills the DBUS_SESSION_BUS_PID

The scripts are called in nepomuk-sandbox-{begin,end}.sh