New Application - NepomukCleaner
The NepomukCleaner application is geared towards removing legacy data
present in the data and cleaning incorrect data. The aim is that we can
drop support for legacy data in the Nepomuk code base, and make it
simpler to read/execute/understand.

For now - It just has one simple job to remove the crappy inferencer
graphs, which was handled by the storage service.