Backup Gui: Remove all the restoration related code
The Restoration is performed by the BackupManager, and the gui is just
supposed to be a rough ui which communicates via dbus. And helps with
the merge conflicts.

This commit breaks the GUI, and it is no longer function. It just
compiles, but then it never used to work properly, so it isn't a big
deal.