Revert "Port to resultiterator"
This reverts commit 6d245db57a02d0a59a71224f0f404796d7fc82c2.

Unfortunately, the result iterator does not allow us to set a custom
model, and we do not want to expose that in the ResultIterator as it is
too low level.

This was not a problem when the query service was independent of the
storage service, but now that it has been merged, we want it to directly
use the model instead of communicating via the NepomukMainModel which
uses a local socket.

Conflicts:
	services/storage/query/searchrunnable.cpp
	services/storage/query/searchrunnable.h