NepomukCleaner: Akonadi migration job
With KDE 4.9, we made sure that all nepomuk resources have a uri of the
form nepomuk:/res/uuid. The new akonadi indexers follow this, but there
is still a lot of legacy data which follows the 'akonadi:?item=<num>'.
This migration job changes all of them to the proper uri and sets the
akonadi uri as the nie:url.