kinotify: Do not assert on receiving an unmount event for an empty path
On unmounting certain usb devices, we get an unmount event from inotify
with and empty name but a valid watch descriptor.

Normally, we would have been able to translate it to a name, but since
we remove all the watches before unmounting it does not exist. Therefore
we must avoid crashing.