Storage: Overwrite the ResourceManager model
Currently the QueryService internally uses the ResourceWatcher and the
QueryParser, both of which use the Resource class. The Resource Class
internally, requires the ResourceManager which uses the
NepomukMainModel. This main model uses a local socket to
connect to the storage service.

This extra socket communication is not required since we can directly
access the model.

In order to do this we to make sure the ResourceManager always returns
the overwrite model, when it has been set.

REVIEW: 106318