FileIndexer: Poppler Based Extractor
Added a very simple extractor for PDF files based on poppler. The
indexing is currently done in two parts. The first part is just the
basic metadata via the SimpleIndexer, and the second part is the file
specific metadata.

We have a simple Extractor class, from which the Extractors are
derived. Eventually, the architecture will need to be improved, but for
this this will have to work.