FileIndexer: Remove threading code from the IndexScheduler
Since most of the work of the IndexScheduling work is done by a separate
process, there is no need for it to run in a separate thread.

The only slight amount of blocking work that is done is listing the
contents of a directory and testing if they should be indexed. And that
is fairly fast in practise.