FileIndexer: Use two queues when indexing files
We have two queues - Slow and Fast. The Fast one is just responsible for
indexing the basic content of the file - stat results + mime type. The
slow one actually analyzes the contents of the file.

The patch comments out a large part of the current functionality. It
will be restored in time.