Use Virtuoso's inbuilt prefix expansion instead of the NRLModel
The NRL Model is terribly slow and consumes as high as 60% of the actual
time it takes to run some queries. Arguably, these queries were small
and virtuoso ran them quite fast. But it still is a lot of unnecessary
time spent on modifying the query. (Based on callgrind)

Since virtuoso already needs to parse the sparql query, it seems to make
more sense to let it do the job. Plus, it cleans up our code rather
nicely.

BUG: 304361
REVIEW: 106474