FileIndexer: Rename and queues and split the work properly
There are now two queues - BasicIndexingQueue, and FileIndexingQueue.

The BasicIndexing queue is responsible for iterating over all the
indexed directories and looking for a file which it has not indexed. It
then indexes the basic metadata of the file - url + mimetype + stat
result

The FileIndexingQueue is responsible for properly indexing all the files
indexed by the BasicIndexingQueue by extracting the actual metadata from
the file.