NepomukIndexer: Add a special debug mode
In the debug mode, the contents of the file are first cleared, the
simple indexer is run and then the file indexer is run. The first 2
steps are generally performed in the fileindexing service, but they are
done here for debugging purposes.

In the normal mode, only the file extractor plugins are called, and the
graph is saved.