ResourceMerger: Literal range check optimization
When checking the literal range, we have a special case for xsd:duration
since Soprano does not support it. During that check we compare both
the Soprano::LiteralTerm's data type (should be ungsigned int) and we
compare the range to the duration xml schema uri.

Comparing the data type is a LOT faster, so we should do that first.
That way we can occasionally save the expensive url comparsion check.

Micro Optimization++