Sync::ResourceIdentifier: Remove unused parts
The original design of the Resource Identifier was way over the top, and
was meant to satisfy a number of different use cases. Currently, it is
just used in storeResources.

Throwing away the unused parts -

* Optional Properties - Aren't used in identification. They are just
  stored in a list that is never used.

* Ignore Resource - A convenient way of removing all the statements
  related to a particular resource uri

* Force Resource - Identify a uri with a particular Resource (with a
  nie:url), and accordingly update the nie:url of all the unidentified
  uris which may be in the same folder or sub folder/files of that
  particular Resource

The ForceResource function may still be useful in the backup file
resolution, but it does not need to exist over here.