StoreResources: Minor optimization by removing unnecessary identification statements
For resolved nodes, we used to add the extra type + nie:url for
identification. This is not required, as the nie:url has been provided
and that is enough for correctly identifying the resource.

Also, added some missing ontologies. They don't really cause any tests
to fail, but it's nice to have them.