StoreResources: Add a flag to force duplicate detection in the graph
By default each SimpleResource in the graph was always hash (an
expensive process) and then checked for duplicates with the other
SimpleResources in the graph.

This feature was only added cause the PIM guys were pushing large
quantities of duplicate data. It doesn't make sense for everyone to pay
the penalty for one application.

They can enable this feature with the MergeDuplicateResources flag.

REVIEW: 106711