ResourceMerger: Refactor to use SyncResource instead of Statements
In StoreResources, the ResourceIdentifier uses a QHash<KUrl, SyncResource>
to manage its data. This provides a convenient high level representation
of Resources, and allows easy data manipulation.

The ResourceMerger, in contrast, was based on Statements. This resulted
in us having to group together various statements for type deduction and
metadata checks. Also, we could not optimize on a higher level, because
we had to deal with each statement individually.

The ResourceMerger now uses a ResourceHash (QHash<KUrl, SyncResource>)
in its main merge function. This simplifies the code a lot, and will
allow us to perform many optimizations in future commits.

This port internally still uses Statements when pushing the data. In
converts the SyncResources to statements.