ResourceMerger: Properly inform the ResourceWatcher Manager
We were still relying on the earlier method of informing the
ResourceWatcher which was based on organizing all the statements into a
series of nested hash tables.

Now that we have moved to SyncResource, we can just rely on
SyncResource's internal hash table. This greatly simplifies the code.

Also fix the broken test - testStoreResources_propertyChanged - It was
trying to access the signal spy out of bounds. This was not caught
earlier cause the order to the signals has now been slightly changed.