ResourceMerger: Improve blank node resolution and creation
In the ResourceMerger, when blank nodes were being resolved they were
created in the model at that instant. Additionally, the resHash keys
were not updated.

Now, when resolving blank nodes, they are simply assigned a nepomuk uri
and the required metaproperties are added to the ResHash. This way, all
the statement additions happen in one place.