ResourceMerger: Optimize insertion and removal of duplicate statements
The duplicate statements are the ones that already exist in the
repository but under a different graph. They need to be given special
treatement, as the graphs need to be merged and all.

After all of this, the previous statements must be removed and the new
ones must be added. That part is has now been optimized. This has been
done by deleting + inserting all the statements in each graph in one go.

It reduces the overhead of calling virtuoso, and we can construct the
query ourselves.

Also fix testRemoveDataByApplication_nieUrl - It inserted statements
with no graph.