FileIndexer: Return the correct file/folder in the IndexScheduler
We have two queues now, so we aren't exactly sure which file/folder to
return. Plus, we cannot always return the flags, as only one of the
queues has them.

So, for now we only return the url and flags from the basic indexing
queue. The other queue will never be user visible cause it is only
activated when the user is idle.