StoreResources: New flag - OverwriteAllProperties
This allows to to use StoreResources as a convenient way to replace
large quantities of data into Nepomuk.

This is especially useful for feeders such as telepathy that do not
follow the traditional - removeDataByApp + storeResources. It just uses
storeResources to update the few properties that usually change.

In that case it makes it easier to just rely on storeResources.

REVIEW: 106770