When we get an EventUnmount with path '/', do not break the while loop.
Breaking means that we lose any further events in the buffer,
and that unmounted folders are not correctly removed from
the path hashes.

Instead, fix the crash by changing the assert to just not trigger if path is / and the event is EventUnmount.
Then do not emit unmounted with a meaningless path.

This fixes a potential bug before it was released anywhere

REVIEW: 107259