Revert "Make KDEPIM 4.6 compile with master"
This reverts commit 0089ee64beb14244f4970b083782766e3bc3a0c2.

This commit was done because of changes in the ontology which changed
the maxCardinality of various properties. This lead to the resource
generator generating setProperty( Type t ) instead of
setProperty( QList<T> t ).

We're reverting the commit as we have moved to Nepomuk, and rcgen2 and
this hack to generate both setProperty( Type t ) and setProperty(
QList<T> t).