New Duplicate Contact cleaning job
There are often a large number of contacts in the nepomuk repository of
identical nco:Contacts. It makes sense to merge them.

We consider two nco:Contact to be identical if they share the same first
name, and have no other literal properties other than the first name. In
the future, maybe we can loosen up the critera, but for now this serve
as a nice way of cleaning up legacy data created before the
pre-storeResources/strigi-feeder times.