NepomukMainModel: Never reset the localSocketModel
We now never create a new LocalSocketModel on reconnecting. Instead we
continue using the old one. This way we do not leak memory, as we could
not delete the old model (open iterators).

This is only possibly cause of the recent changes in Soprano which never
delete the inner ClientConnection which is shared between the model and
the LocalSocketClient. Before that commit the ClientConnection was
deleted which would result in the old model becoming invalid.

REVIEW: 107533