Move the InvalidFileResourceCleaner to the NepomukCleaner
The InvalidFileResourceCleaner was a good job to have. However, running
it automatically on startup caused more problems than it solved. For the
cases when the removeable media handling didn't work, or when mounting
something that Solid did not recognize, it would delete those file
resources on restart (if the device wasn't mounted)

Additionally, it caused extra network usage when mounting a network file
system, by stating (QFile::exists) all the files that had metadata. This
would annoy a lot of users.

We now have it as a job in the nepomukcleaner, which can be run when the
user wants. Ideally one should also have a warning in place that it
might delete metadata for non-mounted media. But we are in string
freeze.

Another reason this makes sense is cause the kioslaves now do not show
any resource which cannot be stated, so even if they are some file
resources which have an incorrect url, they will not be shown to the
user.

BUG: 303659