FileWatch: Remove any existing dirIterator when removing the watches
When issuing a command to remove the watches, watches could be being
added at that point. In that case one should remove those iterators so
that watches are not added after it is unmounted. Also, the dirIterators
seem to prevent the medium from being unmounted.

Also make sure _k_addWatches is invoked if the dir iterater list is not
empty. There can be a case when there is only 1 dir iterator, and it
fails to add a watch. In that case _k_addWatches is not called again,
and that iterator is never removed. This happened to me for a
'lost+found' directory.

BUG: 304943
FIXED-IN: 4.10